requirements.txt
Flask
Pillow
python-escpos
Flask-CORS
pyserial
